# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_logic3579_optimus_internal_infra_response.Envelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      message_key:
        type: string
    type: object
...
  internal_modules_credentials_sshkey.CreateRequest:
    properties:
      description:
        maxLength: 4096
        type: string
      name:
        maxLength: 128
        type: string
      passphrase:
        type: string
      private_key:
        type: string
      username:
        maxLength: 64
        type: string
    required:
    - name
    - private_key
    - username
    type: object
  internal_modules_credentials_sshkey.UpdateRequest:
    properties:
      description:
        maxLength: 4096
        type: string
      name:
        maxLength: 128
        type: string
      passphrase:
        type: string
      private_key:
        type: string
      username:
        maxLength: 64
        type: string
    type: object
...
  internal_modules_k8s_cluster.CreateRequest:
    properties:
      context:
        maxLength: 128
        type: string
      description:
        maxLength: 4096
        type: string
      kubeconfig_id:
        type: integer
      name:
        maxLength: 64
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - context
    - kubeconfig_id
    - name
    type: object
  internal_modules_k8s_cluster.UpdateRequest:
    properties:
      context:
        maxLength: 128
        type: string
      description:
        maxLength: 4096
        type: string
      kubeconfig_id:
        type: integer
      name:
        maxLength: 64
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    P0 admin backend for Optimus — auth, RBAC, users, roles, permissions, menus, audit.
    All authenticated endpoints expect `Authorization: Bearer <access_token>`.
  title: Optimus Admin API
  version: "1.0"
paths:
...
  /credentials/ssh-keys:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: page size (default 20)
        in: query
        name: page_size
        type: integer
      - description: search by name or description
        in: query
        name: q
        type: string
      - description: filter by username (exact)
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: List SSH credentials
      tags:
      - credentials
    post:
      consumes:
      - application/json
      parameters:
      - description: credential payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_credentials_sshkey.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create SSH credential
      tags:
      - credentials
  /credentials/ssh-keys/{id}:
    delete:
      parameters:
      - description: credential ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete SSH credential
      tags:
      - credentials
    get:
      parameters:
      - description: credential ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get SSH credential
      tags:
      - credentials
    put:
      consumes:
      - application/json
      parameters:
      - description: credential ID
        in: path
        name: id
        required: true
        type: integer
      - description: credential payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_credentials_sshkey.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update SSH credential
      tags:
      - credentials
...
  /k8s/clusters:
    get:
      parameters:
      - description: page (default 1)
        in: query
        name: page
        type: integer
      - description: page size (default 20)
        in: query
        name: page_size
        type: integer
      - description: name / description substring
        in: query
        name: search
        type: string
      - description: tag containment filter (charset [a-zA-Z0-9_.-])
        in: query
        name: tag
        type: string
      - description: filter by kubeconfig id
        in: query
        name: kubeconfig_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: List clusters
      tags:
      - k8s
    post:
      consumes:
      - application/json
      parameters:
      - description: cluster payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_k8s_cluster.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create cluster
      tags:
      - k8s
  /k8s/clusters/{id}:
    delete:
      parameters:
      - description: cluster ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Delete cluster
      tags:
      - k8s
    get:
      parameters:
      - description: cluster ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get cluster
      tags:
      - k8s
    put:
      consumes:
      - application/json
      parameters:
      - description: cluster ID
        in: path
        name: id
        required: true
        type: integer
      - description: cluster payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_k8s_cluster.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_logic3579_optimus_internal_infra_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update cluster
      tags:
      - k8s
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"